+++ src/backend/Storage/ExprCalc.Storage/Resources/DatabaseManagement/IDatabaseController.cs
@@ -26,5 +26,6 @@ namespace ExprCalc.Storage.Resources.DatabaseManagement
 
         Task<int> ResetNonFinalStateToPendingAsync(DateTime maxCreatedAt, DateTime newUpdatedAt, CancellationToken token);
         Task<int> DeleteCalculationsAsync(DateTime createdBefore, CancellationToken token);
+        Task<bool> DeleteCalculationByIdAsync(Guid id, CancellationToken token);
     }
 }